pyproject.toml
[project]
name = "c2526-r3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ee>=0.2",
    "geemap>=0.37.1",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pyproj>=3.7.2",
    "python-dotenv>=1.2.1",
    "rasterio>=1.5.0",
    "requests>=2.32.5",
    "timedelta>=2020.12.3",
]
